The Producer and Consumer code duplication fixed
