+++ Dockerfile
@@ -3,11 +3,11 @@ FROM python:3.11-slim
 WORKDIR /app
 
 # Install dependencies
-COPY ../Project/I-Drill/requirements.txt .
+COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copy entire project structure
-COPY ../Project/I-Drill .
+COPY . .
 
 # Set default environment variables
 ENV PYTHONPATH=/app
