+++ src/backend/app.py
@@ -1,17 +1,18 @@
-# Sure! Here's a friendly and professional message you can send to your team worker:
+from fastapi import FastAPI
+from backend.app import test_postgres_connection
+from backend.app import test_influx_connection, write_sample_point
 
-# Hello and Warm Greetings!  
+app = FastAPI()
 
-# Hi Team Member's  
+@app.get("/health")
+def health_check():
+    return {"status": "ok"}
 
-# I hope this message finds you well! I just wanted to take a moment to say hello and check in with you. You’ve been doing an amazing job, and I truly appreciate all the effort and dedication you bring to our team.  
+@app.get("/test/postgres")
+def test_postgres():
+    return {"postgres_connected": test_postgres_connection()}
 
-# If there’s anything you need—whether it’s support, resources, or just a quick chat—please don’t hesitate to let me know. Let’s continue working together to achieve great things!  
-
-# Looking forward to collaborating more in the days ahead.  
-
-# Best regards,  
-# farid kaki  
-
-
-# Let me know if you'd like to personalize it further! 😊
+@app.get("/test/influx")
+def test_influx():
+    ok = write_sample_point()
+    return {"influx_connected": ok and test_influx_connection()}
